Create loss module
