+++ src/otx/algo/visual_prompting/losses/__init__.py
@@ -1,0 +1,8 @@
+# Copyright (C) 2024 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+#
+"""Loss modules for OTX visual prompting model."""
+
+from .sam_loss import SAMCriterion
+
+__all__ = ["SAMCriterion"]
